demo0: Rename dir default -> demo0
Everything that is related to demoing and/or testing the functionality
of iommutests should be in tests/demo0 and should be turned off by
default so demo0 tests are not generated nor compiled. We leave it on
by default with the expectation of "turning it off" when iommutests
grows a new directory under tests.

Signed-off-by: Joel Granados <joel.granados@kernel.org>